% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendDTR.R
\name{recommendDTR}
\alias{recommendDTR}
\title{Recommend the optimal treatment at each stage}
\usage{
recommendDTR(DTRs, X.new, A.new = NULL, Y.new = NULL, start = 1)
}
\arguments{
\item{DTRs}{Output from \code{\link[=learnDTR]{learnDTR()}} that belong to class \code{metaDTR}.}

\item{X.new}{A list of covariates at each stage. In practice, since recommendations are made
stage by stage, here one stage of \code{X.new} will provide one stage's recommendation of action.}

\item{A.new}{A list of observed or predicted actions (from the algorithm). Users can provide real actions observed
rather than algorithm recommended to make this function more flexible.
It allows to handle cases that the realized action/observation may not be consistent
with algorithm results. Only needed when \code{all.inclusive = TRUE}.
Default is \code{NULL}.}

\item{Y.new}{A list. Only required if \code{all.inclusive = TRUE} in \code{\link[=learnDTR]{learnDTR()}}. Default is \code{NULL}}

\item{start}{It allows predictions to start from middle stages (to save time),
in which case \code{A.new} and \code{Y.new}
should be properly provided if \code{all.inclusive = TRUE}.}
}
\value{
It returns the optimal action/treatment recommendations at each stage and results are stored in a list for
each meta-learner method.
}
\description{
This function make recommendations of optimal treatment for a given subject at given stage
}
\details{
This function make recommendations based on the trained learners from \code{\link[=learnDTR]{learnDTR()}} for new dataset.
}
\seealso{
\code{\link{learnDTR}}
}
\author{
Junyi Zhou \email{junyzhou@iu.edu}
}
