% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendDTR.cont.R
\name{recommendDTR.cont}
\alias{recommendDTR.cont}
\title{Recommend the optimal treatment at each stage (Continuous Treatment)}
\usage{
recommendDTR.cont(
  DTRs,
  currentDTRs = NULL,
  X.new,
  A.new = NULL,
  Y.new = NULL,
  A.feasible = NULL,
  A.cnstr.func = NULL,
  n.grid = 100,
  parallel = FALSE,
  parallel.package = NULL,
  verbose = NULL
)
}
\arguments{
\item{DTRs}{Output from \code{\link[=learnDTR.cont]{learnDTR.cont()}} that belong to class \code{metaDTR}.}

\item{currentDTRs}{Object from \code{\link[=recommendDTR.cont]{recommendDTR.cont()}}. Records the results from previous stages so that users can
flexibly make predictions stage by stage. See details and example.}

\item{X.new}{A list of covariates at each stage. In practice, since recommendations are made
stage by stage, here one stage of \code{X.new} will provide one stage's recommendation of action.}

\item{A.new}{A list of observed or predicted actions (from the algorithm). Users can provide real actions observed
rather than algorithm recommended to make this function more flexible.
It allows to handle cases that the realized action/observation may not be consistent
with algorithm results. Only needed when \code{all.inclusive = TRUE}.
Default is \code{NULL}.}

\item{Y.new}{A list. Required if \code{include.Y > 0} in \code{\link[=learnDTR.cont]{learnDTR.cont()}}. Default is \code{NULL}}

\item{A.feasible}{Optional list, default is \code{NULL}. This allow user to specify the subject level's feasible action/treatment
space. The length of list should be equal to the number of stages, and each element should be
an N x 2 of matrix, where N represents the number of subjects and each row is the range of
feasible action/treatment, i.e., (min, max). If \code{A.cnstr.func} is specified in \code{\link[=learnDTR.cont]{learnDTR.cont()}},
then feasible set will also be made to satisfy that.}

\item{A.cnstr.func}{Same as in \code{\link[=learnDTR.cont]{learnDTR.cont()}}. Provide a chance for it to be override, not very common though.
If not specified, it will be inherited directly from the returns of \code{\link[=learnDTR.cont]{learnDTR.cont()}}.}

\item{n.grid}{Same as in \code{\link[=learnDTR.cont]{learnDTR.cont()}}. If not specified, it will be inherited from \code{\link[=learnDTR.cont]{learnDTR.cont()}}.}

\item{parallel}{A boolean, for whether parallel computing is adopted. Also, if a numeric value, it implies the
number of cores to use. Otherwise, directly use the number from \code{\link[=learnDTR.cont]{learnDTR.cont()}}}

\item{parallel.package}{One of c("doMC", "snow", "doParallel"), the parallel package to use.}

\item{verbose}{Console output allowed? Default is \code{NULL}, which will inherit the argument input of}
}
\value{
It returns the optimal action/treatment recommendations at each stage and results are stored in a list for
each meta-learner method.
}
\description{
This function make recommendations of optimal treatment for a given subject at given stage
}
\details{
This function make recommendations based on the trained learners from \code{\link[=learnDTR.cont]{learnDTR.cont()}} for new dataset.
Since in real application, later stage covariates/outcomes are unobservable until treatments are
assigned. So in most cases, \code{\link[=recommendDTR.cont]{recommendDTR.cont()}} needs to be applied stage by stage, which is allowed in
this package. User can provide available information stage by stage, and obtain the optimal recommendations
at each stage, iteratively.
}
\examples{
## Similar example as recommendDTR()
DTRs = learnDTR.cont(X = ThreeStg_Dat$X,
                     A = ThreeStg_Dat$A,
                     Y = ThreeStg_Dat$Y,
                     weights = rep(1, 3),
                     baseLearner  = c("RF"),
                     metaLearners = c("S"),
                     include.X = 1,
                     include.A = 2,
                     include.Y = 0)
optDTR <- recommendDTR.cont(DTRs, currentDTRs = NULL,
                            X.new = ThreeStg_Dat$X.test)
}
\seealso{
\code{\link{learnDTR.cont}}
}
\author{
Junyi Zhou \email{junyzhou@iu.edu}
}
